requirements.txt
flask
pandas
numpy
joblib
scikit-learn
python-dotenv
requests
groq
